+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ### 1.1.1
 * added the log viewer and the changelogs viewer;
 * many buttons are disabled after the click, to prevent some actions are performed repeatedly;
+* usernames and the name of user groups cannot be changed. Improved permissions about users and user groups;
 * added the changelog file.